Improve names of SqlClientPool implementations
