+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/pool/ReactiveConnection.java
@@ -22,7 +22,7 @@ import java.util.concurrent.CompletionStage;
  * operation must be chained on completion of the first operation.
  * This restriction might be relaxed in future, and is due to the
  * implementation of the {@code ProxyConnection} returned by
- * {@link org.hibernate.reactive.pool.impl.SqlClientPool#getProxyConnection()}.
+ * {@link org.hibernate.reactive.pool.impl.DefaultSqlClientPool#getProxyConnection()}.
  *
  * @see ReactiveConnectionPool
  */
