+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/configuration/JdbcUrlParserTest.java
@@ -5,7 +5,7 @@
  */
 package org.hibernate.reactive.configuration;
 
-import org.hibernate.reactive.pool.impl.SqlClientPool;
+import org.hibernate.reactive.pool.impl.DefaultSqlClientPool;
 import org.junit.Test;
 
 import java.net.URI;
@@ -16,31 +16,31 @@ public class JdbcUrlParserTest {
 
 	@Test
 	public void returnsNullForNull() {
-		URI uri = SqlClientPool.parse( null );
+		URI uri = DefaultSqlClientPool.parse( null );
 		assertThat( uri ).isNull();
 	}
 
 	@Test
 	public void uriCreation() {
-		URI uri = SqlClientPool.parse("jdbc:postgresql://localhost:5432/hreact");
+		URI uri = DefaultSqlClientPool.parse( "jdbc:postgresql://localhost:5432/hreact");
 		assertThat(uri).isNotNull();
 	}
 
 	@Test
 	public void parsePort() {
-		URI uri = SqlClientPool.parse("jdbc:postgresql://localhost:5432/hreact");
+		URI uri = DefaultSqlClientPool.parse( "jdbc:postgresql://localhost:5432/hreact");
 		assertThat(uri).hasPort(5432);
 	}
 
 	@Test
 	public void parseHost() {
-		URI uri = SqlClientPool.parse("jdbc:postgresql://localhost:5432/hreact");
+		URI uri = DefaultSqlClientPool.parse( "jdbc:postgresql://localhost:5432/hreact");
 		assertThat(uri).hasHost("localhost");
 	}
 
 	@Test
 	public void parseScheme() {
-		URI uri = SqlClientPool.parse("jdbc:postgresql://localhost:5432/hreact");
+		URI uri = DefaultSqlClientPool.parse( "jdbc:postgresql://localhost:5432/hreact");
 		assertThat(uri).hasScheme("postgresql");
 	}
 }
